Rename 'clients' to 'integrations'
