+++ src/infrastructure/integrations/contracts.go
@@ -1,0 +1,4 @@
+package integrations
+
+type IHttpClient interface {
+}
