+++ src/infrastructure/integrations/sendgrid_email_client.go
@@ -1,4 +1,4 @@
-package clients
+package integrations
 
 import (
 	"fmt"
